Add integration test to test places search api Move test data to data access project

+++ src/Itinerary.Business/Places/Models/Place.cs
@@ -9,7 +9,7 @@ namespace Itinerary.Business.Places.Models
       Categories = new List<Category>();
     }
 
-    public int PlaceId { get; set; }
+    public long PlaceId { get; set; }
 
     public string Name { get; set; }
